Florida 70 Precinct Shape File

This data set contains the 70 Precinct shapefile and related data for each precinct.

data("fl70")

Format

sf data.frame containing columns for useful data related to the redistricting process, subsetted from real data in Florida, and sf geometry column.

geoid

Contains unique identifier for each precinct which can be matched to the full Florida dataset.

pop

Contains the population of each precinct.

vap

Contains the voting age population of each precinct.

obama

Contains the 2012 presidential vote for Obama.

mccain

Contains the 2012 presidential vote for McCain.

TotPop

Contains the population of each precinct. Identical to pop.

BlackPop

Contains the black population of each precinct.

HispPop

Contains the Hispanic population of each precinct.

VAP

Contains the voting age population of each precinct. Identical to vap.

BlackVAP

Contains the voting age population of black constituents of each precinct.

HispVAP

Contains the voting age population of hispanic constituents of each precinct.

geometry

Contains sf geometry of each precinct.

Details

It is a random 70 precinct connected subset from Florida's precincts. This was introduced by <doi:10.1080/2330443X.2020.1791773>

References

Benjamin Fifield, Kosuke Imai, Jun Kawahara & Christopher T. Kenny (2020) The Essential Role of Empirical Validation in Legislative Redistricting Simulation, Statistics and Public Policy, 7:1, 52-68, doi:10.1080/2330443X.2020.1791773

Examples

data(fl70)