Flip Constraints Helper

Creates a constraints list for use with redist_flip.

flip_constraints_helper(
  map,
  constraint = "compact",
  constraintweight = 0.6,
  init_plan = NULL,
  compactness_metric = "edges-removed",
  areas,
  borderlength_mat,
  ssdmat,
  ssd_denom,
  counties = NULL,
  partisan_metric = "efficiency-gap",
  rvote,
  dvote,
  group_pop = NULL,
  target_min = 0.55,
  target_other = 0.25,
  minorityprop
)

Arguments

map
A redist_map object.

constraint
character vector of constraints to use. Current accepted are "compact", "population", "countysplit", "hinge", "vra", "minority", "similarity", "partisan", and "segregation." Defaults to compact.

constraintweight
corresponding weights to use with constraint. Weights must be nonzero if provided. Defaults to a weak compactness constraint

init_plan
initial plan to use for the similarity constraint

compactness_metric
character with "edges-removed", "polsby-popper", or "fryer-holden". Default is edges-removed.

areas
areas to use with compact:polsby-popper. Computed from map if not provided and needed.

borderlength_mat
border lengths to use with compact:polsby-popper. Computed from map if not provided and needed.

ssdmat
squared distance matrix to use with compact:fryer-holden. Computed from map if not provided and needed.

ssd_denom
Denominator for compact:fryer-holden. Defaults to 1.0

counties
tidy evaluated column with counties in map.

partisan_metric
character with "proportional-representation" or "efficiency-gap"

rvote
numeric vector of votes for Republicans. used with partisan constraint.

dvote
numeric vector of votes for Democrats used with partisan constraint.

group_pop
tidy evaluated column with group populations in map. Used with hinge, segregation, vra, and minority.

target_min
numeric with higher minority target for vra. Defaults to 0.55.

target_other
numeric with higher minority target for vra. Defaults to 0.25.

minorityprop
minority proportions to target. Used by minority and hinge.Defaults to c(0.55) if none specified.

Value

list with an entry for each constraint usable with redist_flip()

Examples

data(iowa)
iowa_map <- redist_map(iowa, existing_plan = cd_2010, total_pop = pop, pop_tol = 0.01)
cons <- flip_constraints_helper(
  map = iowa_map, constraint = c('compact', 'vra'),
  constraintweight = c(0.05, 10), target_min = 0.05,
  target_other = 0.01, group_pop = black
)

redist_flip(iowa_map, 10, constraints = cons)

#> 
#> ==================== 
#> redist.flip(): Automated Redistricting Simulation Using
#>          Markov Chain Monte Carlo
#> 
#> Preprocessing data.
#> 
#> Starting swMH().
#> 10 percent done.
#> Metropolis acceptance ratio: inf
#> 
#> 20 percent done.
#> Metropolis acceptance ratio: 2
#> 
#> 30 percent done.
#> Metropolis acceptance ratio: 1.5
#> 
#> 40 percent done.
#> Metropolis acceptance ratio: 1.33333
#> 
#> 50 percent done.
#> Metropolis acceptance ratio: 1.25
#> 
#> 60 percent done.
#> Metropolis acceptance ratio: 1.2
#> 
#> 70 percent done.
#> Metropolis acceptance ratio: 1.16667
#> 
#> 80 percent done.
#> Metropolis acceptance ratio: 1.14286
#> 
#> 90 percent done.
#> Metropolis acceptance ratio: 1.125
#> 
#> 100 percent done.
#> Metropolis acceptance ratio: 1.11111
#> 

#> 10 sampled plans and 1 reference plan with 4 districts from a 99-unit map,
#>   drawn using an unknown algorithm
#> Plans matrix: num [1:99, 1:11] 3 3 1 2 4 1 1 4 1 1 ...
#> # A tibble: 44 × 14
#>    draw    district total_pop distance_parity distance_origin… mhdecisions mhprob
#>    <fct>      <int>     <dbl>           <dbl>            <dbl>       <dbl>  <dbl>
#>  1 cd_2010        1    761548        NA                 NA              NA NA    
#>  2 cd_2010        2    761624        NA                 NA              NA NA    
#>  3 cd_2010        3    761612        NA                 NA              NA NA    
#>  4 cd_2010        4    761571        NA                 NA              NA NA    
#>  5 1              1    761548         0.00678            0.990           1  0.952
#>  6 1              2    766755         0.00678            0.990           1  0.952
#>  7 1              3    756481         0.00678            0.990           1  0.952
#>  8 1              4    761571         0.00678            0.990           1  0.952
#>  9 2              1    761548         0.00806            0.980           1  1.00 
#> 10 2              2    766755         0.00806            0.980           1  1.00 
#> # … with 34 more rows, and 7 more variables: pparam <dbl>, beta_sequence <dbl>,
#> #   energy_psi <dbl>, boundary_partitions <dbl>, boundary_ratio <dbl>,
#> #   constraint_compact <dbl>, constraint_vra <dbl>