redist: Simulation Methods for Legislative Redistricting

Enables researchers to sample redistricting plans from a pre-specified target distribution using Sequential Monte Carlo and Markov Chain Monte Carlo algorithms. The package allows for the implementation of various constraints in the redistricting process such as geographic compactness and population parity requirements. Tools for analysis such as computation of various summary statistics and plotting functionality are also included. The package implements the SMC algorithm of McCartan and Imai (2023) doi:10.1214/23-AOAS1763 , the enumeration algorithm of Fifield, Imai, Kawahara, and Kenny (2020) doi:10.1080/2330443X.2020.1791773 , the Flip MCMC algorithm of Fifield, Higgins, Imai and Tarr (2020) doi:10.1080/10618600.2020.1739532 , the Merge-split/Recombination algorithms of Carter et al. (2019) arXiv:1911.01503 and DeFord et al. (2021) doi:10.1162/99608f92.eb30390f , and the Short-burst optimization algorithm of Cannon et al. (2020) arXiv:2011.02288.

See also

Author

Maintainer: Christopher T. Kenny christopherkenny@fas.harvard.edu

Authors:

Cory McCartan corymccartan@nyu.edu

Ben Fifield benfifield@gmail.com

Kosuke Imai imai@harvard.edu

Other contributors:

Jun Kawahara jkawahara@i.kyoto-u.ac.jp [contributor]

Alexander Tarr atarr@princeton.edu [contributor]

Michael Higgins mikehiggins@k-state.edu [contributor]