Run parameter testing for redist.flip

redist.findparams is used to find optimal parameter values of redist.flip for a given map.

redist.findparams(
  adj,
  total_pop,
  nsims,
  ndists = NULL,
  init_plan = NULL,
  adapt_lambda = FALSE,
  adapt_eprob = FALSE,
  params,
  ssdmat = NULL,
  group_pop = NULL,
  counties = NULL,
  nstartval_store = 1,
  maxdist_startval = 100,
  maxiterrsg = 5000,
  report_all = TRUE,
  parallel = FALSE,
  ncores = NULL,
  log = FALSE,
  verbose = TRUE
)

Arguments

adj
An adjacency matrix, list, or object of class "SpatialPolygonsDataFrame."

total_pop
A vector containing the populations of each geographic unit.

nsims
The number of simulations run before a save point.

ndists
The number of congressional districts. The default is NULL.

init_plan
A vector containing the congressional district labels of each geographic unit. The default is NULL. If not provided, random and contiguous congressional district assignments will be generated using redist.rsg.

adapt_lambda
Whether to adaptively tune the lambda parameter so that the Metropolis-Hastings acceptance probability falls between 20% and 40%. Default is FALSE.

adapt_eprob
Whether to adaptively tune the edgecut probability parameter so that the Metropolis-Hastings acceptance probability falls between 20% and 40%. Default is FALSE.

params
A matrix of parameter values to test, such as the output of expand.grid. Parameters accepted for params include eprob, lambda, pop_tol, beta, and constraint.

ssdmat
A matrix of squared distances between geographic units. The default is NULL.

group_pop
A vector of populations for some sub-group of interest. The default is NULL.

counties
A vector of county membership assignments. The default is NULL.

nstartval_store
The number of maps to sample from the preprocessing chain for use as starting values in future simulations. Default is 1.

maxdist_startval
The maximum distance from the starting map that sampled maps should be. Default is 100 (no restriction).

maxiterrsg
Maximum number of iterations for random seed-and-grow algorithm to generate starting values. Default is 5000.

report_all
Whether to report all summary statistics for each set of parameter values. Default is TRUE.

parallel
Whether to run separate parameter settings in parallel. Default is FALSE.

ncores
Number of parallel tasks to run, declared outside of the function. Default is NULL.

log
Whether to open a log to track progress for each parameter combination being tested. Default is FALSE.

verbose
Whether to print additional information about the tests. Default is TRUE.

Value

redist.findparams returns a print-out of summary statistics about each parameter setting.

Details

This function allows users to test multiple parameter settings of redist.flip in preparation for a longer run for analysis.

References

Fifield, Benjamin, Michael Higgins, Kosuke Imai and Alexander Tarr. (2016) "A New Automated Redistricting Simulator Using Markov Chain Monte Carlo." Working Paper. Available at http://imai.princeton.edu/research/files/redist.pdf.

Examples

# \donttest{
data(fl25)
data(fl25_enum)
data(fl25_adj)

## Get an initial partition
init_plan <- fl25_enum$plans[, 5118]

params <- expand.grid(eprob = c(.01, .05, .1))

## Run the algorithm
redist.findparams(adj = fl25_adj, total_pop = fl25$pop,
                  init_plan = init_plan, nsims = 10000, params = params)

#> ## ------------------------------
#>  ## redist.findparams(): Parameter tuning for redist.flip()
#>  ## Searching over 3 parameter combinations
#>  ## ------------------------------
#> 
#> 
#> ==================== 
#> redist.flip(): Automated Redistricting Simulation Using
#>          Markov Chain Monte Carlo
#> 
#> Preprocessing data.
#> 
#> Starting swMH().
#> 10 percent done.
#> Metropolis acceptance ratio: 0.986987
#> 
#> 20 percent done.
#> Metropolis acceptance ratio: 0.990495
#> 
#> 30 percent done.
#> Metropolis acceptance ratio: 0.991997
#> 
#> 40 percent done.
#> Metropolis acceptance ratio: 0.991748
#> 
#> 50 percent done.
#> Metropolis acceptance ratio: 0.992599
#> 
#> 60 percent done.
#> Metropolis acceptance ratio: 0.992332
#> 
#> 70 percent done.
#> Metropolis acceptance ratio: 0.992142
#> 
#> 80 percent done.
#> Metropolis acceptance ratio: 0.992374
#> 
#> 90 percent done.
#> Metropolis acceptance ratio: 0.992221
#> 
#> 100 percent done.
#> Metropolis acceptance ratio: 0.992499
#> 
#> 
#> ==================== 
#> redist.flip(): Automated Redistricting Simulation Using
#>          Markov Chain Monte Carlo
#> 
#> Preprocessing data.
#> 
#> Starting swMH().
#> 10 percent done.
#> Metropolis acceptance ratio: 0.965966
#> 
#> 20 percent done.
#> Metropolis acceptance ratio: 0.966483
#> 
#> 30 percent done.
#> Metropolis acceptance ratio: 0.961654
#> 
#> 40 percent done.
#> Metropolis acceptance ratio: 0.965491
#> 
#> 50 percent done.
#> Metropolis acceptance ratio: 0.965793
#> 
#> 60 percent done.
#> Metropolis acceptance ratio: 0.965328
#> 
#> 70 percent done.
#> Metropolis acceptance ratio: 0.964995
#> 
#> 80 percent done.
#> Metropolis acceptance ratio: 0.96362
#> 
#> 90 percent done.
#> Metropolis acceptance ratio: 0.963107
#> 
#> 100 percent done.
#> Metropolis acceptance ratio: 0.963396
#> 
#> 
#> ==================== 
#> redist.flip(): Automated Redistricting Simulation Using
#>          Markov Chain Monte Carlo
#> 
#> Preprocessing data.
#> 
#> Starting swMH().
#> 10 percent done.
#> Metropolis acceptance ratio: 0.930931
#> 
#> 20 percent done.
#> Metropolis acceptance ratio: 0.923462
#> 
#> 30 percent done.
#> Metropolis acceptance ratio: 0.919973
#> 
#> 40 percent done.
#> Metropolis acceptance ratio: 0.924231
#> 
#> 50 percent done.
#> Metropolis acceptance ratio: 0.923785
#> 
#> 60 percent done.
#> Metropolis acceptance ratio: 0.924654
#> 
#> 70 percent done.
#> Metropolis acceptance ratio: 0.927132
#> 
#> 80 percent done.
#> Metropolis acceptance ratio: 0.927366
#> 
#> 90 percent done.
#> Metropolis acceptance ratio: 0.926214
#> 
#> 100 percent done.
#> Metropolis acceptance ratio: 0.926493
#> 
#>  ## -------------------------------------
#>  ## -------------------------------------
#>  ## Parameter Values for Simulation 1 
#> ## Edgecut probability = 0.01
#> ## Lambda = 0
#> ## No hard population constraint applied
#> ## Not setting any soft constraints
#> ## -------------------------------------
#> ## Diagnostics:
#> ## Metropolis-Hastings Acceptance Ratio = 0.992
#> ## Mean population parity distance = 0.826
#> ## Median population parity distance = 0.828
#> ## Population parity range = 0.012 1.922
#> ## MCMC Iteration quantiles of population parity median = 0.836 0.906 0.664 0.868
#> 
#> ## Mean share of geographies equal to initial assignment = 0.305
#> ## Median share of geographies equal to initial assignment = 0.28
#> ## Range of share of geographies equal to initial assignment = 0 1
#> ## MCMC Iteration quantiles of geography distance to initial assignment = 0.28 0.32 0.28 0.28
#> ## -------------------------------------
#> ## -------------------------------------
#> 
#>  ## -------------------------------------
#>  ## -------------------------------------
#>  ## Parameter Values for Simulation 1 
#> ## Edgecut probability = 0.05
#> ## Lambda = 0
#> ## No hard population constraint applied
#> ## Not setting any soft constraints
#> ## -------------------------------------
#> ## Diagnostics:
#> ## Metropolis-Hastings Acceptance Ratio = 0.963
#> ## Mean population parity distance = 0.892
#> ## Median population parity distance = 0.882
#> ## Population parity range = 0.021 1.921
#> ## MCMC Iteration quantiles of population parity median = 0.899 0.894 0.868 0.876
#> 
#> ## Mean share of geographies equal to initial assignment = 0.33
#> ## Median share of geographies equal to initial assignment = 0.32
#> ## Range of share of geographies equal to initial assignment = 0 0.96
#> ## MCMC Iteration quantiles of geography distance to initial assignment = 0.4 0.32 0.24 0.28
#> ## -------------------------------------
#> ## -------------------------------------
#> 
#>  ## -------------------------------------
#>  ## -------------------------------------
#>  ## Parameter Values for Simulation 1 
#> ## Edgecut probability = 0.1
#> ## Lambda = 0
#> ## No hard population constraint applied
#> ## Not setting any soft constraints
#> ## -------------------------------------
#> ## Diagnostics:
#> ## Metropolis-Hastings Acceptance Ratio = 0.926
#> ## Mean population parity distance = 0.879
#> ## Median population parity distance = 0.875
#> ## Population parity range = 0.017 1.956
#> ## MCMC Iteration quantiles of population parity median = 0.801 0.869 0.875 0.909
#> 
#> ## Mean share of geographies equal to initial assignment = 0.383
#> ## Median share of geographies equal to initial assignment = 0.36
#> ## Range of share of geographies equal to initial assignment = 0 1
#> ## MCMC Iteration quantiles of geography distance to initial assignment = 0.52 0.28 0.4 0.36
#> ## -------------------------------------
#> ## -------------------------------------
#> 

#> $diagnostics
#> [1] " ## -------------------------------------\n ## -------------------------------------\n ## Parameter Values for Simulation 1 \n## Edgecut probability = 0.01\n## Lambda = 0\n## No hard population constraint applied\n## Not setting any soft constraints\n## -------------------------------------\n## Diagnostics:\n## Metropolis-Hastings Acceptance Ratio = 0.992\n## Mean population parity distance = 0.826\n## Median population parity distance = 0.828\n## Population parity range = 0.012 1.922\n## MCMC Iteration quantiles of population parity median = 0.836 0.906 0.664 0.868\n\n## Mean share of geographies equal to initial assignment = 0.305\n## Median share of geographies equal to initial assignment = 0.28\n## Range of share of geographies equal to initial assignment = 0 1\n## MCMC Iteration quantiles of geography distance to initial assignment = 0.28 0.32 0.28 0.28\n## -------------------------------------\n## -------------------------------------\n\n ## -------------------------------------\n ## -------------------------------------\n ## Parameter Values for Simulation 1 \n## Edgecut probability = 0.05\n## Lambda = 0\n## No hard population constraint applied\n## Not setting any soft constraints\n## -------------------------------------\n## Diagnostics:\n## Metropolis-Hastings Acceptance Ratio = 0.963\n## Mean population parity distance = 0.892\n## Median population parity distance = 0.882\n## Population parity range = 0.021 1.921\n## MCMC Iteration quantiles of population parity median = 0.899 0.894 0.868 0.876\n\n## Mean share of geographies equal to initial assignment = 0.33\n## Median share of geographies equal to initial assignment = 0.32\n## Range of share of geographies equal to initial assignment = 0 0.96\n## MCMC Iteration quantiles of geography distance to initial assignment = 0.4 0.32 0.24 0.28\n## -------------------------------------\n## -------------------------------------\n\n ## -------------------------------------\n ## -------------------------------------\n ## Parameter Values for Simulation 1 \n## Edgecut probability = 0.1\n## Lambda = 0\n## No hard population constraint applied\n## Not setting any soft constraints\n## -------------------------------------\n## Diagnostics:\n## Metropolis-Hastings Acceptance Ratio = 0.926\n## Mean population parity distance = 0.879\n## Median population parity distance = 0.875\n## Population parity range = 0.017 1.956\n## MCMC Iteration quantiles of population parity median = 0.801 0.869 0.875 0.909\n\n## Mean share of geographies equal to initial assignment = 0.383\n## Median share of geographies equal to initial assignment = 0.36\n## Range of share of geographies equal to initial assignment = 0 1\n## MCMC Iteration quantiles of geography distance to initial assignment = 0.52 0.28 0.4 0.36\n## -------------------------------------\n## -------------------------------------\n\n"
#> 
#> $startvals
#> $startvals[[1]]
#>       [,1]
#>  [1,]    2
#>  [2,]    2
#>  [3,]    2
#>  [4,]    1
#>  [5,]    2
#>  [6,]    1
#>  [7,]    3
#>  [8,]    3
#>  [9,]    1
#> [10,]    3
#> [11,]    3
#> [12,]    3
#> [13,]    3
#> [14,]    1
#> [15,]    2
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> 
#> $startvals[[2]]
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    3
#>  [4,]    3
#>  [5,]    2
#>  [6,]    3
#>  [7,]    2
#>  [8,]    2
#>  [9,]    3
#> [10,]    2
#> [11,]    1
#> [12,]    3
#> [13,]    3
#> [14,]    3
#> [15,]    3
#> [16,]    1
#> [17,]    1
#> [18,]    3
#> [19,]    1
#> [20,]    1
#> [21,]    3
#> [22,]    1
#> [23,]    3
#> [24,]    1
#> [25,]    1
#> 
#> $startvals[[3]]
#>       [,1]
#>  [1,]    2
#>  [2,]    3
#>  [3,]    3
#>  [4,]    3
#>  [5,]    2
#>  [6,]    1
#>  [7,]    2
#>  [8,]    2
#>  [9,]    1
#> [10,]    2
#> [11,]    2
#> [12,]    2
#> [13,]    2
#> [14,]    2
#> [15,]    2
#> [16,]    3
#> [17,]    3
#> [18,]    3
#> [19,]    2
#> [20,]    2
#> [21,]    1
#> [22,]    2
#> [23,]    1
#> [24,]    2
#> [25,]    2
#> 
#> 

# }