Tidy Flip Markov Chain Monte Carlo

Tidy Flip Markov Chain Monte Carlo

redist_flip(
  map,
  nsims,
  init_plan,
  counties = NULL,
  group_pop,
  constraints = list(),
  eprob = 0.05,
  lambda = 0,
  temper = FALSE,
  betaseq = "powerlaw",
  betaseqlength = 10,
  betaweights = NULL,
  adapt_lambda = FALSE,
  adapt_eprob = FALSE,
  exact_mh = FALSE,
  adjswaps = TRUE,
  init_name = NULL,
  verbose = TRUE
)

Arguments

map
A redist_map object.

nsims
The number of samples to draw.

init_plan
A vector containing the congressional district labels of each geographic unit. The default is NULL. If not provided, a random initial plan will be generated using redist.smc.

counties
A column in map containing county membership

group_pop
A column in map containing group populations

constraints
a list of constraints to implement. Can be created with flip_constraints_helper

eprob
The probability of keeping an edge connected. The default is 0.05.

lambda
lambda The parameter determining the number of swaps to attempt each iteration of the algorihtm. The number of swaps each iteration is equal to Pois(lambda) + 1. The default is 0.

temper
Whether to use simulated tempering algorithm. Default is FALSE.

betaseq
Sequence of beta values for tempering. The default is powerlaw (see Fifield et. al (2020) for details).

betaseqlength
@param betaseqlength Length of beta sequence desired for tempering. The default is 10.

betaweights
betaweights Sequence of weights for different values of beta. Allows the user to upweight certain values of beta over others. The default is NULL (equal weighting).

adapt_lambda
adapt_lambda Whether to adaptively tune the lambda parameter so that the Metropolis-Hastings acceptance probability falls between 20% and 40%. Default is FALSE.

adapt_eprob
eprob Whether to adaptively tune the edgecut probability parameter so that the Metropolis-Hastings acceptance probability falls between 20% and 40%. Default is FALSE.

exact_mh
Whether to use the approximate (FALSE) or exact (TRUE) Metropolis-Hastings ratio calculation for accept-reject rule. Default is FALSE.

adjswaps
Flag to restrict swaps of beta so that only values adjacent to current constraint are proposed. The default is TRUE.

init_name
a name for the initial plan, or FALSE to not include the initial plan in the output. Defaults to the column name of the existing plan, or "<init>" if the initial plan is sampled.

verbose
Whether to print initialization statement. Default is TRUE.

Value

redist plans object

Examples

if (FALSE) {
data(fl25)
fl25_map <- redist_map(fl, ndists = 4, existing_plan = cd)
sims <- redist_flip(map = fl25_map, nsims = 10)
}