A set of redistricting plans

A redist_plans object is essentially a data frame of summary information on each district and each plan, along with the matrix of district assignments and information about the simulation process used to generate the plans.

redist_plans(plans, map, algorithm, wgt = NULL, ...)

Arguments

plans
a matrix with n_precinct columns and n_sims rows, or a single vector of precinct assignments.

map
a redist_map object

algorithm
the algorithm used to generate the plans (usually "smc" or "mcmc")

wgt
the weights to use, if any.

...
Other named attributes to set

Value

a new redist_plans object.

Details

The first two columns of the data frame will be draw, a factor indexing the simulation draw, and district, an integer indexing the districts within a plan. The data frame will therefore have n_sims*ndists rows. As a data frame, the usual dplyr methods will work.

Other useful methods for redist_plans objects: * add_reference * pullback * number_by * match_numbers * prec_assignment * get_plan_matrix * get_plan_weights * get_sampling_info * subset_sampled * subset_ref * as.matrix.redist_plans * plot.redist_plans

Examples

data(iowa)

iowa = redist_map(iowa, existing_plan=cd, pop_tol=0.05)
rsg_plan = redist.rsg(iowa$adj, iowa$pop, ndists=4, pop_tol=0.05)$plan

#> 
#> ==================== 
#> redist.rsg(): Automated Redistricting Starts
#> 
#> 
#> 	4 districts built using 99 precincts in 0.03 seconds...
#> 

redist_plans(rsg_plan, iowa, "rsg")

#> 1 sampled plan with 4 districts from a 99-unit map,
#>   drawn using random seed-and-grow
#> Plans matrix: int [1:99, 1] 3 2 1 2 3 1 1 1 4 1 ...
#> # A tibble: 4 x 3
#>   draw  district total_pop
#> * <fct>    <int>     <dbl>
#> 1 1            1    735800
#> 2 1            2    745063
#> 3 1            3    780796
#> 4 1            4    784696