SMC Redistricting Sampler

redist_smc uses a Sequential Monte Carlo algorithm to generate nearly independent congressional or legislative redistricting plans according to contiguity, population, compactness, and administrative boundary constraints.

redist_smc(
  map,
  nsims,
  counties = NULL,
  compactness = 1,
  constraints = list(),
  resample = TRUE,
  constraint_fn = function(m) rep(0, ncol(m)),
  init_particles = NULL,
  n_steps = NULL,
  adapt_k_thresh = 0.975,
  seq_alpha = 0.2 + 0.3 * compactness,
  truncate = (compactness != 1),
  trunc_fn = redist_quantile_trunc,
  pop_temper = 0,
  final_infl = 1,
  ref_name = NULL,
  verbose = TRUE,
  silent = FALSE
)

Arguments

map
A redist_map object.

nsims
The number of samples to draw.

counties
A vector containing county (or other administrative or geographic unit) labels for each unit, which may be integers ranging from 1 to the number of counties, or a factor or character vector. If provided, the algorithm will only generate maps which split up to ndists-1 counties. If no county-split constraint is desired, this parameter should be left blank.

compactness
Controls the compactness of the generated districts, with higher values preferring more compact districts. Must be nonnegative. See the 'Details' section for more information, and computational considerations.

constraints
A redist_constr() object or a list containing information on sampling constraints. See constraints for more information.

resample
Whether to perform a final resampling step so that the generated plans can be used immediately. Set this to FALSE to perform direct importance sampling estimates, or to adjust the weights manually.

constraint_fn
A function which takes in a matrix where each column is a redistricting plan and outputs a vector of log-weights, which will be added the the final weights.

init_particles
A matrix of partial plans to begin sampling from. For advanced use only. The matrix must have nsims columns and a row for every precinct. It is important to ensure that the existing districts meet contiguity and population constraints, or there may be major issues when sampling.

n_steps
How many steps to run the SMC algorithm for. Each step splits off a new district. Defaults to all remaining districts. If fewer than the number of remaining splits, reference plans are disabled.

adapt_k_thresh
The threshold value used in the heuristic to select a value k_i for each splitting iteration. Set to 0.9999 or 1 if the algorithm does not appear to be sampling from the target distribution. Must be between 0 and 1.

seq_alpha
The amount to adjust the weights by at each resampling step; higher values prefer exploitation, while lower values prefer exploration. Must be between 0 and 1.

truncate
Whether to truncate the importance sampling weights at the final step by trunc_fn. Recommended if compactness is not 1. Truncation only applied if resample=TRUE.

trunc_fn
A function which takes in a vector of weights and returns a truncated vector. If loo package is installed (strongly recommended), will default to Pareto-smoothed Importance Sampling (PSIS) rather than naive truncation.

pop_temper
The strength of the automatic population tempering. Try values of 0.01-0.05 to start if the algorithm gets stuck on the final few splits.

final_infl
A multiplier for the population constraint on the final iteration. Used to loosen the constraint when the sampler is getting stuck on the final split.

ref_name
a name for the existing plan, which will be added as a reference plan, or FALSE to not include the initial plan in the output. Defaults to the column name of the existing plan.

verbose
Whether to print out intermediate information while sampling. Recommended.

silent
Whether to suppress all diagnostic information.

Value

redist_smc returns an object of class redist_plans containing the simulated plans.

Details

This function draws nearly-independent samples from a specific target measure, controlled by the pop_tol, compactness, constraints, and constraint_fn parameters.

Key to ensuring good performance is monitoring the efficiency of the resampling process at each SMC stage. Unless silent=FALSE, this function will print out the effective sample size of each resampling step to allow the user to monitor the efficiency. If verbose=TRUE the function will also print out information on the \(k_i\) values automatically chosen and the acceptance rate (based on the population constraint) at each step.

Higher values of compactness sample more compact districts; setting this parameter to 1 is computationally efficient and generates nicely compact districts. Values of other than 1 may lead to highly variable importance sampling weights. By default these weights are truncated using redist_quantile_trunc to stabilize the resulting estimates, but if truncation is used, a specific truncation function should probably be chosen by the user.

References

McCartan, C., & Imai, K. (2020). Sequential Monte Carlo for Sampling Balanced and Compact Redistricting Plans. Available at https://imai.fas.harvard.edu/research/files/SMCredist.pdf.

Examples

# \donttest{
set.seed(1)
data(fl25)

fl_map = redist_map(fl25, ndists=3, pop_tol=0.1)

#> Projecting to CRS 3857

sampled_basic = redist_smc(fl_map, 10000)

#> SEQUENTIAL MONTE CARLO
#> Sampling 10000 25-unit maps with 3 districts and population between 52512.9 and 64182.4.
#> Making split 1 of 2
#> Using k = 3
#> Iteration 1000/10000
#> Iteration 2000/10000
#> Iteration 3000/10000
#> Iteration 4000/10000
#> Iteration 5000/10000
#> Iteration 6000/10000
#> Iteration 7000/10000
#> Iteration 8000/10000
#> Iteration 9000/10000
#> Iteration 10000/10000
#> 21.6535% acceptance rate.
#> Resampling effective sample size: 9942.23 (99.4223% efficiency).
#> Making split 2 of 2
#> Using k = 2
#> Iteration 1000/10000
#> Iteration 2000/10000
#> Iteration 3000/10000
#> Iteration 4000/10000
#> Iteration 5000/10000
#> Iteration 6000/10000
#> Iteration 7000/10000
#> Iteration 8000/10000
#> Iteration 9000/10000
#> Iteration 10000/10000
#> 6.0808% acceptance rate.
#> Resampling effective sample size: 9696.72 (96.9672% efficiency).

sampled_constr = redist_smc(fl_map, 10000, constraints=list(
                                incumbency = list(strength=100, incumbents=c(3, 6, 25))
                            ))

#> Error: Not a nested list

# }