Sampling constraints

The redist_smc() and redist_mergesplit() algorithms in this package allow for additional constraints on the redistricting process to be encoded in the target distribution for sampling. These functions are provided to specify these constraints. All arguments are quoted and evaluated in the context of the data frame provided to redist_constr().

add_constr_status_quo(constr, strength, current)

add_constr_grp_pow(
  constr,
  strength,
  group_pop,
  total_pop = NULL,
  tgt_group = 0.5,
  tgt_other = 0.5,
  pow = 1
)

add_constr_grp_hinge(
  constr,
  strength,
  group_pop,
  total_pop = NULL,
  tgts_group = c(0.55)
)

add_constr_compet(constr, strength, dvote, rvote, pow = 0.5)

add_constr_incumbency(constr, strength, incumbents)

add_constr_splits(constr, strength)

add_constr_multisplits(constr, strength)

Arguments

strength
The strength of the constraint. Higher values mean a more restrictive constraint.

current
The reference map for the status quo constraint.

group_pop
A vector of group population

total_pop
A vector of total population. Defaults to the population vector used for sampling.

tgt_group, tgt_other
Target group shares for the power-type constraint.

pow
The exponent for the power-type constraint.

tgts_group
A vector of target group shares for the hinge-type constraint.

dvote, rvote
A vector of Democratic or Republican vote counts

incumbents
A vector of precinct indicess for incumbents

Details

All constraints are fed into a Gibbs measure, with coefficients on each constraint set by the corresponding strength parameter. The strength can be any real number, with zero corresponding to no constraint. Higher and higher strength values will eventually cause the algorithm's accuracy and efficiency to suffer. Whenever you use constraints, be sure to check all sampling diagnostics.

The status_quo constraint adds a term measuring the variation of information distance between the plan and the reference, rescaled to [0, 1].

The grp_hinge constraint takes a list of target group percentages. It matches each district to its nearest target percentage, and then applies a penalty of the form \(\sqrt{max(0, tgt - grouppct)}\), summing across districts. This penalizes districts which are below their target percentage.

The grp_pow constraint (for expert use) adds a term of the form \((|tgtgroup-grouppct||tgtother-grouppct|)^{pow})\), which encourages districts to have group shares near either tgt_group or tgt_other.

The compet constraint encourages competitiveness by applying the grp_pow constraint with target percentages set to 50%. For convenience, it is specified with Democratic and Republican vote shares.

The incumbency constraint adds a term counting the number of districts containing paired-up incumbents. Values of strength should generally be small, given that the underlying values are counts.

The splits constraint adds a term counting the number of counties which are split once or more. Values of strength should generally be small, given that the underlying values are counts.

The multisplits constraint adds a term counting the number of counties which are split twice or more. Values of strength should generally be small, given that the underlying values are counts.