Calculate Group Proportion by District

redist.group.percent computes the proportion that a group makes up in each district across a matrix of maps.

group_frac(
  map,
  group_pop,
  total_pop = map[[attr(map, "pop_col")]],
  .data = cur_plans()
)

redist.group.percent(plans, group_pop, total_pop, ncores = 1)

Arguments

map

a redist_map object

group_pop

A numeric vector with the population of the group for every precinct.

total_pop

A numeric vector with the population for every precinct.

.data

a redist_plans object

plans

A matrix with one row for each precinct and one column for each map. Required.

ncores

Number of cores to use for parallel computing. Default is 1.

Value

matrix with percent for each district

Examples

data(fl25)
data(fl25_enum)

cd <- fl25_enum$plans[, fl25_enum$pop_dev <= 0.05]

redist.group.percent(plans = cd,
    group_pop = fl25$BlackPop,
    total_pop = fl25$TotPop)
#>         [,1]    [,2]     [,3]     [,4]     [,5]     [,6]    [,7]     [,8]
#> [1,] 0.15569 0.19637 0.177320 0.177320 0.172667 0.172667 0.16053 0.162539
#> [2,] 0.25248 0.21075 0.258871 0.252477 0.276550 0.284377 0.24574 0.284377
#> [3,] 0.10956 0.10956 0.087466 0.089526 0.069848 0.056922 0.10956 0.071467
#>          [,9]    [,10]    [,11]   [,12]   [,13]   [,14]    [,15]    [,16]
#> [1,] 0.142972 0.182796 0.182796 0.15479 0.15159 0.15870 0.191546 0.191546
#> [2,] 0.298520 0.251830 0.245744 0.25561 0.25894 0.24923 0.073122 0.071467
#> [3,] 0.087466 0.087466 0.089526 0.10956 0.10956 0.10956 0.255612 0.262238
#>         [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#> [1,] 0.175916 0.175916 0.065701 0.065701 0.055238 0.081932 0.081932 0.052479
#> [2,] 0.262238 0.255612 0.260610 0.259451 0.259451 0.241849 0.241020 0.258871
#> [3,] 0.087466 0.089526 0.197628 0.197714 0.204920 0.198648 0.198752 0.204920
#>        [,25]    [,26]    [,27]   [,28]    [,29]    [,30]    [,31]   [,32]
#> [1,] 0.12112 0.051501 0.070945 0.13226 0.171770 0.067677 0.067677 0.15923
#> [2,] 0.28438 0.258937 0.241020 0.27655 0.262238 0.250168 0.249228 0.25017
#> [3,] 0.10956 0.204920 0.204920 0.10956 0.089526 0.197628 0.197714 0.10956
#>         [,33]    [,34]   [,35]    [,36]    [,37]   [,38]    [,39]   [,40]
#> [1,] 0.090424 0.142020 0.15672 0.072963 0.059096 0.14374 0.135781 0.12702
#> [2,] 0.224312 0.284377 0.25183 0.241849 0.250168 0.11922 0.070008 0.19186
#> [3,] 0.204920 0.089526 0.10956 0.204920 0.204920 0.25702 0.313522 0.19865
#>        [,41]   [,42]    [,43]    [,44]    [,45]   [,46]    [,47]    [,48]
#> [1,] 0.12702 0.12302 0.123022 0.057485 0.134208 0.16716 0.150167 0.158729
#> [2,] 0.19186 0.27470 0.082589 0.260610 0.313522 0.11922 0.054231 0.047071
#> [3,] 0.19875 0.11922 0.313522 0.204920 0.069848 0.23658 0.313522 0.313522
#>         [,49]    [,50]   [,51]    [,52]   [,53]   [,54]    [,55]   [,56]
#> [1,] 0.195477 0.178413 0.16995 0.127858 0.11581 0.11581 0.115813 0.11963
#> [2,] 0.070008 0.080975 0.15059 0.313522 0.25507 0.27470 0.087419 0.19875
#> [3,] 0.257022 0.257022 0.19865 0.074049 0.15059 0.12658 0.313522 0.20417
#>         [,57]    [,58]    [,59]    [,60]    [,61]    [,62]    [,63]   [,64]
#> [1,] 0.099131 0.099131 0.144763 0.142555 0.134017 0.117300 0.133355 0.13532
#> [2,] 0.163698 0.186065 0.055697 0.062487 0.071467 0.313522 0.071191 0.12658
#> [3,] 0.257022 0.236579 0.313522 0.313522 0.313522 0.087466 0.313522 0.25702
#>        [,65]   [,66]   [,67]    [,68]   [,69]   [,70]   [,71]    [,72]    [,73]
#> [1,] 0.13532 0.11023 0.17060 0.145246 0.16092 0.17032 0.15805 0.173717 0.130748
#> [2,] 0.15059 0.20417 0.15059 0.056922 0.25702 0.15059 0.25702 0.274702 0.073122
#> [3,] 0.23658 0.20492 0.19763 0.313522 0.10280 0.19875 0.10457 0.071467 0.313522
#>         [,74]   [,75]    [,76]   [,77]    [,78]    [,79]    [,80]    [,81]
#> [1,] 0.155913 0.17097 0.094771 0.16252 0.191966 0.114451 0.238294 0.208378
#> [2,] 0.047848 0.15059 0.313522 0.15059 0.071191 0.313522 0.080975 0.054231
#> [3,] 0.313522 0.19771 0.109556 0.20492 0.257022 0.089526 0.197628 0.257022
#>         [,82]    [,83]    [,84]   [,85]    [,86]    [,87]    [,88]   [,89]
#> [1,] 0.183832 0.201341 0.236634 0.14754 0.192347 0.177133 0.202074 0.20462
#> [2,] 0.257022 0.062487 0.080975 0.05515 0.257022 0.257022 0.056922 0.11922
#> [3,] 0.074049 0.257022 0.198648 0.31352 0.069848 0.087466 0.257022 0.19875
#>         [,90]   [,91]    [,92]   [,93]    [,94]   [,95]    [,96]    [,97]
#> [1,] 0.193517 0.19325 0.200559 0.19228 0.157006 0.18373 0.217294 0.217588
#> [2,] 0.071467 0.12658 0.055697 0.12658 0.274702 0.23658 0.070008 0.047071
#> [3,] 0.257022 0.19763 0.257022 0.19865 0.087466 0.10280 0.236579 0.257022
#>         [,98]    [,99]  [,100]   [,101]   [,102]   [,103]  [,104]   [,105]
#> [1,] 0.260531 0.254680 0.19228 0.188738 0.180280 0.252700 0.22160 0.215317
#> [2,] 0.055697 0.071191 0.12658 0.073122 0.255067 0.071191 0.19875 0.071467
#> [3,] 0.197628 0.197628 0.19875 0.257022 0.087466 0.198648 0.10280 0.236579
#>       [,106]   [,107]   [,108]  [,109]   [,110]  [,111]   [,112]   [,113]
#> [1,] 0.14906 0.213492 0.261208 0.18550 0.199492 0.20472 0.269151 0.213696
#> [2,] 0.16370 0.071191 0.056922 0.12658 0.236579 0.05515 0.054231 0.236579
#> [3,] 0.20492 0.236579 0.198648 0.20492 0.087466 0.25702 0.198648 0.069848
#>        [,114]  [,115]   [,116]  [,117]   [,118]   [,119]   [,120]   [,121]
#> [1,] 0.235920 0.21801 0.241570 0.18053 0.222050 0.255232 0.256347 0.279312
#> [2,] 0.080975 0.19865 0.198648 0.23658 0.069149 0.071467 0.070008 0.047071
#> [3,] 0.198752 0.10457 0.074049 0.10457 0.236579 0.198648 0.198752 0.198648
#>       [,122]   [,123]   [,124]   [,125]   [,126]   [,127]   [,128]   [,129]
#> [1,] 0.19324 0.153093 0.169385 0.238522 0.243288 0.196412 0.229331 0.252595
#> [2,] 0.12658 0.274702 0.274702 0.047071 0.197628 0.255067 0.054231 0.198648
#> [3,] 0.19771 0.089526 0.073122 0.236579 0.074049 0.071467 0.236579 0.069848
#>        [,130]  [,131]   [,132]  [,133]   [,134]   [,135]  [,136]   [,137]
#> [1,] 0.237543 0.23977 0.263258 0.13197 0.222708 0.262525 0.21944 0.222681
#> [2,] 0.080975 0.05238 0.056922 0.27470 0.056922 0.062487 0.19763 0.062487
#> [3,] 0.197714 0.23658 0.197628 0.10956 0.236579 0.198648 0.10457 0.236579
#>       [,138]   [,139]   [,140]   [,141]   [,142]   [,143]   [,144]   [,145]
#> [1,] 0.19521 0.172758 0.258560 0.230694 0.254556 0.213716 0.271380 0.175888
#> [2,] 0.11902 0.257022 0.055697 0.080975 0.197628 0.047848 0.054231 0.255067
#> [3,] 0.20492 0.089526 0.198648 0.204920 0.069848 0.257022 0.197628 0.089526
#>        [,146]  [,147]   [,148]   [,149]   [,150]   [,151]   [,152]   [,153]
#> [1,] 0.251596 0.22541 0.242463 0.254165 0.280205 0.240785 0.242396 0.276500
#> [2,] 0.198752 0.05515 0.197714 0.071467 0.047071 0.198752 0.069349 0.047848
#> [3,] 0.069848 0.23658 0.074049 0.198752 0.197714 0.074049 0.204920 0.197628
#>        [,154]   [,155]   [,156]   [,157]   [,158]   [,159]   [,160]   [,161]
#> [1,] 0.259430 0.191629 0.194680 0.251692 0.267875 0.263438 0.234389 0.261348
#> [2,] 0.055697 0.255067 0.236579 0.071191 0.054231 0.062487 0.047848 0.062487
#> [3,] 0.197714 0.073122 0.089526 0.198752 0.198752 0.197714 0.236579 0.198752
#>        [,162]  [,163]   [,164]   [,165]   [,166]   [,167]  [,168]   [,169]
#> [1,] 0.270047 0.26429 0.235789 0.210129 0.224806 0.260091 0.15462 0.262091
#> [2,] 0.054231 0.05515 0.074049 0.073122 0.087419 0.056922 0.25507 0.056922
#> [3,] 0.197714 0.19865 0.204920 0.236579 0.204920 0.198752 0.10956 0.197714
#>       [,170]   [,171]   [,172]   [,173]   [,174]   [,175]   [,176]   [,177]
#> [1,] 0.21758 0.253622 0.263679 0.274207 0.253509 0.248728 0.257506 0.277850
#> [2,] 0.19875 0.071191 0.054231 0.047848 0.197714 0.073122 0.055697 0.047071
#> [3,] 0.10457 0.197714 0.204920 0.198648 0.069848 0.198648 0.198752 0.198752
#>        [,178]  [,179]  [,180]   [,181]   [,182]  [,183]   [,184]   [,185]
#> [1,] 0.253120 0.26641 0.21898 0.275090 0.272856 0.26518 0.268874 0.247796
#> [2,] 0.055697 0.05515 0.19771 0.047848 0.047848 0.05515 0.047848 0.073122
#> [3,] 0.204920 0.19763 0.10457 0.197714 0.198752 0.19771 0.204920 0.198752
#>       [,186]  [,187]   [,188]  [,189]   [,190]   [,191]   [,192]
#> [1,] 0.23144 0.25883 0.222882 0.26312 0.233113 0.227284 0.218911
#> [2,] 0.08157 0.05515 0.089298 0.05515 0.081056 0.083029 0.090914
#> [3,] 0.20492 0.20492 0.204920 0.19875 0.204920 0.204920 0.204920